Styling improvements

+++ src/app/components/form.tsx
@@ -2,6 +2,33 @@
 import { useEffect, useState } from "react";
 import { useForm, SubmitHandler, set, useFieldArray } from "react-hook-form";
 import { type Request } from "./models";
+import { getReasonPhrase } from "http-status-codes";
+
+type Response = {
+  status: number;
+  headers: Headers;
+  body: string;
+  time: number;
+  size: string;
+  cookies: string;
+  statusText: string;
+};
+
+function formatFileSize(bytes: number) {
+  const sizes = ["Bytes", "KB", "MB", "GB", "TB"];
+  const decimalDigits = 2;
+
+  if (bytes === 0) {
+    return "0 Bytes";
+  }
+
+  const i = Math.floor(Math.log(bytes) / Math.log(1024));
+  const formattedSize = parseFloat(
+    (bytes / Math.pow(1024, i)).toFixed(decimalDigits)
+  );
+
+  return `${formattedSize} ${sizes[i]}`;
+}
 
 export default function Form({
   activeRequestIndex,
@@ -10,12 +37,12 @@ export default function Form({
   activeRequestIndex: number;
   setActiveRequestIndex: (index: number) => void;
 }) {
+  const [activeTab, setActiveTab] = useState<"data" | "headers">("data");
   const [editAsJson, setEditAsJson] = useState(false);
   const [formDataAsJson, setFormDataAsJson] = useState<null | string>(null);
   const [submitting, setSubmitting] = useState(false);
-  const [response, setResponse] = useState<null | string>(null);
-  const [responseHeaders, setResponseHeaders] = useState<null | string>(null);
-  const [responseStatus, setResponseStatus] = useState<null | string>(null);
+  const [response, setResponse] = useState<null | Response>(null);
+
   const {
     register,
     handleSubmit,
@@ -38,6 +65,7 @@ export default function Form({
   const onSubmit: SubmitHandler<Request> = async (data) => {
     setSubmitting(true);
     console.log(data);
+    const startTime = performance.now();
     try {
       const res = await fetch(data.url, {
         method: data.action,
@@ -45,25 +73,42 @@ export default function Form({
           data.headers.map((header) => [header.key, header.value])
         ),
       });
+      const endTime = performance.now();
+      const requestTime = endTime - startTime;
+      // Get the response size in kilobytes
+      const contentLength = res.headers.get("content-length") ?? "0";
+
       const json = await res.json();
       console.log(json);
       console.log(res);
       setSubmitting(false);
-      setResponse(JSON.stringify(json, null, 2));
-      setResponseHeaders(JSON.stringify(res.headers, null, 2));
-      setResponseStatus(res.status.toString());
+      setResponse({
+        status: res.status,
+        headers: res.headers,
+        body: JSON.stringify(json, null, 2),
+        time: requestTime,
+        size: formatFileSize(parseInt(contentLength, 10)),
+        cookies: res.headers.get("set-cookie") ?? "",
+        statusText: getReasonPhrase(res.status),
+      });
 
       const existingRequests =
         JSON.parse(window.localStorage.getItem("requests") ?? "[]") || [];
+      if (
+        existingRequests &&
+        existingRequests.length > 0 &&
+        JSON.stringify(existingRequests[existingRequests.length - 1]) ===
+          JSON.stringify(data)
+      ) {
+        return;
+      }
       existingRequests.push(data);
       window.localStorage.setItem("requests", JSON.stringify(existingRequests));
       setActiveRequestIndex(existingRequests.length - 1);
     } catch (error) {
       console.log(error);
       setSubmitting(false);
-      setResponse(JSON.stringify(error, null, 2));
-      setResponseHeaders(null);
-      setResponseStatus(null);
+      setResponse(null);
     }
   };
 
@@ -88,6 +133,9 @@ export default function Form({
 
   return (
     <div className="flex flex-col items-start lg:items-center w-full lg:w-1/2 mx-auto p-4  rounded shadow">
+      <div className="prose  w-full">
+        <h2>Request</h2>
+      </div>
       <button
         className="button"
         type="button"
@@ -157,6 +205,27 @@ export default function Form({
             <input
               className="join-item btn"
               type="radio"
+              aria-label="PUT"
+              value="PUT"
+              {...register("action")}
+            />
+            <input
+              className="join-item btn"
+              type="radio"
+              aria-label="DELETE"
+              value="DELETE"
+              {...register("action")}
+            />
+            <input
+              className="join-item btn"
+              type="radio"
+              aria-label="OPTIONS"
+              value="OPTIONS"
+              {...register("action")}
+            />
+            <input
+              className="join-item btn"
+              type="radio"
               aria-label="PATCH"
               value="PATCH"
               {...register("action")}
@@ -173,7 +242,7 @@ export default function Form({
                 })}
                 defaultValue={header.key}
                 placeholder="Key"
-                className="w-full p-2 border border-gray-300 rounded mb-2 lg:mb-0"
+                className="w-full p-2 input input-bordered mb-2 lg:mb-0"
               />
               <input
                 {...register(`headers[${index}].value` as any, {
@@ -181,21 +250,34 @@ export default function Form({
                 })}
                 defaultValue={header.value}
                 placeholder="Value"
-                className="w-full p-2 border border-gray-300 rounded"
+                className="w-full p-2 input input-bordered"
               />
               <button
                 type="button"
                 onClick={() => remove(index)}
-                className="self-start lg:self-center p-2 text-white bg-red-500 hover:bg-red-700 rounded"
+                className="self-start lg:self-center p-2 btn btn-error btn-square btn-outline"
               >
-                X
+                <svg
+                  xmlns="http://www.w3.org/2000/svg"
+                  className="h-6 w-6"
+                  fill="none"
+                  viewBox="0 0 24 24"
+                  stroke="currentColor"
+                >
+                  <path
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    strokeWidth="2"
+                    d="M6 18L18 6M6 6l12 12"
+                  />
+                </svg>
               </button>
             </div>
           ))}
           <button
             type="button"
             onClick={() => append({ key: "", value: "" })}
-            className="w-full p-2 text-white bg-green-500 hover:bg-green-700 rounded mb-4"
+            className="w-full p-2 btn btn-accent mb-4"
           >
             Add New Header
           </button>
@@ -208,36 +290,87 @@ export default function Form({
         </form>
       )}
       <div className="w-full">
-        <div className="form-control mb-4">
-          <label className="label">
-            <span className="label-text">Response Status</span>
-          </label>
-          {responseStatus && (
-            <span className="transition-opacity">{responseStatus ?? ""}</span>
-          )}
+        <div className="divider"></div>
+        <div className="prose ">
+          <h2>Response</h2>
         </div>
         <div className="form-control mb-4">
           <label className="label">
-            <span className="label-text">Response</span>
+            <span className="label-text">Status:</span>
+            <label className="swap  swap-active text-right">
+              <div
+                className={
+                  !response?.status && !submitting ? "swap-on" : "swap-off"
+                }
+              >
+                Pending...
+              </div>
+              <div className={submitting ? "swap-on" : "swap-off"}>
+                <span className="loading loading-ring loading-xs"></span>
+              </div>
+              <div
+                className={
+                  response?.status && !submitting
+                    ? "swap-on font-bold"
+                    : "swap-off"
+                }
+              >
+                {response?.status} - {response?.statusText} -{" "}
+                {Math.floor(response?.time ?? 0)} ms{" "}
+              </div>
+            </label>
           </label>
-          <textarea
-            className="textarea h-52 textarea-bordered"
-            placeholder="Type here"
-            value={response ?? ""}
-            readOnly
-          ></textarea>
         </div>
-        <div className="form-control mb-4">
-          <label className="label">
-            <span className="label-text">Response Headers</span>
-          </label>
-          <textarea
-            className="textarea h-24 textarea-bordered textarea-lg"
-            placeholder="Type here"
-            value={responseHeaders ?? ""}
-            readOnly
-          ></textarea>
+        <div className="w-full tabs">
+          <a
+            className={
+              activeTab === "data"
+                ? "tab tab-bordered tab-active"
+                : "tab tab-bordered"
+            }
+            onClick={() => setActiveTab("data")}
+          >
+            Data
+          </a>
+          <a
+            className={
+              activeTab === "headers"
+                ? "tab tab-bordered tab-active"
+                : "tab tab-bordered"
+            }
+            onClick={() => setActiveTab("headers")}
+          >
+            Headers
+          </a>
         </div>
+        {activeTab === "data" && (
+          <div className="form-control mb-4">
+            <label className="label">
+              <span className="label-text">Response Data</span>
+            </label>
+            <textarea
+              className="textarea h-52 textarea-bordered"
+              placeholder="Waiting for response..."
+              value={response?.body ?? ""}
+              readOnly
+            ></textarea>
+          </div>
+        )}
+        {activeTab === "headers" && (
+          <div className="form-control mb-4">
+            <label className="label">
+              <span className="label-text">Response Headers</span>
+            </label>
+            {Array.from(response?.headers.entries() ?? []).map(
+              ([key, value]) => (
+                <div key={key} className="flex flex-row">
+                  <div className="w-1/2">{key}</div>
+                  <div className="w-1/2">{value}</div>
+                </div>
+              )
+            )}
+          </div>
+        )}
       </div>
     </div>
   );
